backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy 
psycopg2-binary
alembic
python-dotenv==1.0.1
passlib[bcrypt]
openai==1.30.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
aiohttp

azure-search-documents
azure-ai-ml
azure-storage-blob
azure-core
